+++ Flutter/lib/Core/txplayer_define.dart
@@ -435,6 +435,25 @@ abstract class TXVodPlayEvent {
   static const EXTRA_SUBTITLE_START_POSITION_MS = "startPositionMs";
   static const EXTRA_SUBTITLE_DURATION_MS = "durationMs";
   static const EXTRA_SUBTITLE_TRACK_INDEX = "trackIndex";
+
+  /// Alternative playback URL for HEVC downgrade playback, supported by the Advanced Player 12.0
+  /// HEVC 降级播放时备选播放 URL， 播放器高级版 12.0 版本开始支持
+  static const VOD_KEY_BACKUP_URL = "VOD_KEY_BACKUP_URL";
+  /// When HEVC is downgraded, the main video encoding type
+  /// HEVC 降级播放时，主播放的视频编码类型
+  static const VOD_KEY_VIDEO_CODEC_TYPE = "VOD_KEY_VIDEO_CODEC_TYPE";
+  /// MediaType of alternative playback URL resource during HEVC downgrade playback, supported by the Advanced Player 12.0
+  /// HEVC 降级播放时备选播放 URL 资源的 MediaType， 播放器高级版 12.0 版本开始支持
+  static const VOD_KEY_BACKUP_URL_MEDIA_TYPE = "VOD_KEY_BACKUP_URL_MEDIA_TYPE";
+  /// HEVC format, supported by the player advanced version 12.0
+  /// HEVC 格式， 播放器高级版 12.0 版本开始支持
+  static const VOD_PLAY_MIMETYPE_H265 = "video/hevc";
+  /// mp4加密播放：不加密。 12.2 版本开始支持。
+  /// MP4 encryption playback: No encryption. Supported since version 12.2.
+  static const MP4_ENCRYPTION_LEVEL_NONE = 0;
+  /// mp4加密播放： mp4本地加密播放。12.2 版本开始支持。
+  /// MP4 encrypted playback: MP4 local encrypted playback. Supported since version 12.2.
+  static const MP4_ENCRYPTION_LEVEL_L2 = 2;
 }
 
 abstract class TXVodNetEvent {
@@ -618,8 +637,9 @@ class TXPlayInfoParams {
   // video url, only applicable for preloading. When using it, you only need to fill in either the url or fileId.
   // The priority of the url is higher than that of the fileId.
   final String? url;
-
-  const TXPlayInfoParams({required this.appId, required this.fileId, this.psign = "", this.url = ""});
+  // Custom httpHeader
+  final Map<String, String>? httpHeader;
+  const TXPlayInfoParams({required this.appId, required this.fileId, this.psign = "", this.url = "", this.httpHeader});
 
   Map<String, dynamic> toJson() {
     Map<String, dynamic> json = {};
@@ -627,6 +647,7 @@ class TXPlayInfoParams {
     json["fileId"] = fileId;
     json["psign"] = psign;
     json["url"] = url;
+    json["httpHeader"] = httpHeader;
     return json;
   }
 }