+++ Flutter/lib/Core/txplayer_messages.dart
@@ -381,6 +381,7 @@ class FTXVodPlayConfigPlayerMsg {
     this.enableRenderProcess,
     this.preferredResolution,
     this.mediaType,
+    this.encryptedMp4Level,
   });
 
   int? playerId;
@@ -427,6 +428,9 @@ class FTXVodPlayConfigPlayerMsg {
   /// 媒资类型，默认auto类型, 取值参考 see[TXVodPlayEvent]
   int? mediaType;
 
+  /// MP4 encryption playback:  Supported since version 12.2.
+  int? encryptedMp4Level;
+
   Object encode() {
     return <Object?>[
       playerId,
@@ -450,6 +454,7 @@ class FTXVodPlayConfigPlayerMsg {
       enableRenderProcess,
       preferredResolution,
       mediaType,
+      encryptedMp4Level,
     ];
   }
 
@@ -477,6 +482,7 @@ class FTXVodPlayConfigPlayerMsg {
       enableRenderProcess: result[18] as bool?,
       preferredResolution: result[19] as int?,
       mediaType: result[20] as int?,
+      encryptedMp4Level: result[21] as int?,
     );
   }
 }
@@ -885,6 +891,7 @@ class PreLoadInfoMsg {
     this.preloadSizeMB,
     this.preferredResolution,
     this.tmpPreloadTaskId,
+    this.httpHeader,
   });
 
   int? appId;
@@ -901,6 +908,8 @@ class PreLoadInfoMsg {
 
   int? tmpPreloadTaskId;
 
+  Map<String?, String?>? httpHeader;
+
   Object encode() {
     return <Object?>[
       appId,
@@ -910,6 +919,7 @@ class PreLoadInfoMsg {
       preloadSizeMB,
       preferredResolution,
       tmpPreloadTaskId,
+      httpHeader,
     ];
   }
 
@@ -923,6 +933,7 @@ class PreLoadInfoMsg {
       preloadSizeMB: result[4] as double?,
       preferredResolution: result[5] as int?,
       tmpPreloadTaskId: result[6] as int?,
+      httpHeader: (result[7] as Map<Object?, Object?>?)?.cast<String?, String?>(),
     );
   }
 }