+++ Flutter/CHANGELOG.md
@@ -76,3 +76,14 @@ Player_Premium 分支：Android & iOS 端集成TXLiteAVSDK_Player_Premium lastes
 - Fix the issue of reversed logic in the live streaming mute method.
 - iOS adds support for Picture-in-Picture for live streaming, which requires premium permission to use.
 - Fix other known issues.
+
+#### Version: 12.2.0 2024.12.04
+
+##### Features：
+
+- set Android TXLiteAVSDK_Player_Premium to 12.2.0.15065，tag：release_premium_v12.2.0
+- set iOS TXLiteAVSDK_Player_Premium to 12.2.16945， tag：release_premium_v12.2.0
+- Pre-download supports httpHeader
+- Supports encrypted playback of MP4
+- Added support for HEVC playback downgrade
+- Fix other known issues.